Addressing comments from the bunny
